Remove insantiation of mappers
I couldn't think of a real use case for it.

I also want to find a better way of working with mappers. The DataMapper
guys are doing something like `DataMapper[Article]` which is a great way
to go between mappers and mapped classes. That's not to say I want to
clone their idea, but it got me thinking that `Article` and
`ArticleMapper` makes sense to humans reasoning about it, but I had to
do an ugly hack to get the Mapper class to pair mappers and mapped classes.