Use MongoDB cursor's limit method rather than arg
This is a huge optimization for indexed queries. If you sort with the
limit argument, you're still comparing all documents in the collection.
Sorting with the cursor's limit method, you only look at the N documents
you're limiting to.

For example, if you have 1000 documents in an `Article` collection,
which has a descending-order index on the `timestamp` attribute, and you
execute the following query:

```javascript
db.Article.find({}).sort({timestamp: -1}).limit(10)
```

You will only scan the 10 objects you want in the results. It can do
this because it knows the top 10 results based entirely on the index.

Before this commit, we were executing this query:

```javascript
db.Article.find({}, {limit: 10}).sort({timestamp: -1})
```

This was comparing all 1000 documents in the `Article` collection. Found
this while running an `explain()` on both queries.