Fix sorting results from MongoDB
Sorting got messed up in the conversion to Moped and, for some reason,
there wasn't an integration test for that. This commit fixes both of
those problems.

Also, trivial but whatever, `MongoDB#sort` used to call the object
passed in `cursor` because that's what the Mongo gem called it, but I
changed the name to `query` because it's a damn query. :-)