Remove non-mapper-specified indexes from the DB
This commit has `Mapper#reindex!` completely synchronize the indexes in
the mapper with the indexes in the DB. This means any indexes in the
mapper that are not in the DB will be created and any indexes not in the
mapper that are in the DB will be deleted. Mapper indexes are canon.