Switch Rbx version and remove JRuby from Travis
The old Rubinius version is no longer supported on Travis.

JRuby won't build at all because we test allow testing against the
Postgres adapter, which depends on the `pg` gem which doesn't work on
JRuby due to the C extension. Once we find a way to get that adapter
working on JRuby, I'll add it back into the CI matrix.