Remove updating object in memory with `update`
This was a poorly thought-out feature and doesn't work well with
serialization. I'm not sure whether the feature is a good or bad idea at
the moment but it's current implementation, at least, is not good, so
it's getting removed.

Besides, users should be using the `save` method to save their objects.
`update` is an implementation detail.