nil id does not indicate persistence
Previously, we were relying on the idea that the mapper was the only
thing that defined the id instance variable. This works fine until
anything else defines that ivar, like setting an accessor to it with
Virtus.

Ideally, a model detached from the database wouldn't care about the id
at all and that's what I relied on when I wrote that code. However, we
don't live in an ideal world. Sometimes people need to be able to access
this ivar — for example, mapping to URLs in a web app. For this reason,
we can no longer rely on the ideal case.